requirements.txt
torch
transformers
datasets
peft
accelerate
safetensors
tokenizers
bitsandbytes
sentencepiece
matplotlib
tqdm
numpy
faiss-cpu